backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
gunicorn
whitenoise
python-dotenv
dj-database-url
psycopg2-binary